Revert "Simplify differences between play commands"
