--- run_simple.bat
@@ -1,5 +1,0 @@
-@echo off
-
-node index.js
-
-pause
